my_ml_project/requirements.txt
asgiref
Django
joblib
numpy
opencv-python
pillow
pip
pywavelets
scikit-learn
scipy
sqlparse
threadpoolctl
tzdata
